Upgrade @metamask/snaps-controllers to v21
Upgrades @metamask/snaps-controllers from v20 to v21. The new version
moves snap lifecycle event tracking (install, update, uninstall, etc.)
into the controller itself via an AnalyticsController:trackEvent action,
so the subscriptions that previously lived in MetaMaskController are
removed. The trackEvent hook on SnapController is similarly removed. v21
also improves metrics for preinstalled OTA updates.

Extracts getClientConfig() into a shared utils.ts to avoid the version
parsing logic being duplicated across SnapController and
SnapRegistryController init files, and adds tests for it.
