+++ app/scripts/messenger-client-init/messengers/snaps/snap-controller-messenger.ts
@@ -44,9 +44,15 @@ export function getSnapControllerMessenger(
       'ExecutionService:outboundRequest',
       'ExecutionService:outboundResponse',
       'KeyringController:lock',
+      // 'SnapController:snapInstallStarted',
+      // 'SnapController:snapInstallFailed',
+      // 'SnapController:snapInstalled',
+      // 'SnapController:snapUpdated',
+      // 'SnapController:snapUninstalled',
       'SnapRegistryController:registryUpdated',
     ],
     actions: [
+      'AnalyticsController:trackEvent',
       'PermissionController:getEndowments',
       'PermissionController:getPermissions',
       'PermissionController:hasPermission',
