+++ package.json
@@ -223,7 +223,7 @@
     "@types/readable-stream-3@^4.0.4": "npm:@types/readable-stream@^4.0.4",
     "readable-stream-2@^2.3.3": "npm:readable-stream@^2.3.3",
     "readable-stream-3@^3.6.2": "npm:readable-stream@^3.6.2",
-    "@metamask/snaps-sdk": "^11.1.0",
+    "@metamask/snaps-sdk": "^11.1.1",
     "@babel/core": "7.29.7",
     "@babel/runtime": "patch:@babel/runtime@npm%3A7.29.2#~/.yarn/patches/@babel-runtime-npm-7.29.2-b49cad1c67.patch",
     "@spruceid/siwe-parser@npm:2.1.0": "patch:@spruceid/siwe-parser@npm%3A2.1.0#~/.yarn/patches/@spruceid-siwe-parser-npm-2.1.0-060b7ede7a.patch",
@@ -301,7 +301,7 @@
     "tmp": "^0.2.7",
     "postcss-loader/jiti": "npm:npm-empty-package@1.0.0",
     "esbuild": "0.28.1",
-    "@metamask/snaps-controllers": "^20.0.5",
+    "@metamask/snaps-controllers": "^21.0.0",
     "@metamask/permission-controller@npm:^12.2.0": "^13.0.0",
     "@metamask/permission-controller@npm:^12.2.1": "^13.0.0",
     "@metamask/permission-controller@npm:^12.3.0": "^13.0.0",
@@ -436,7 +436,7 @@
     "@metamask/signature-controller": "^39.2.6",
     "@metamask/smart-transactions-controller": "^24.2.2",
     "@metamask/snap-account-service": "^1.0.0",
-    "@metamask/snaps-controllers": "^20.0.5",
+    "@metamask/snaps-controllers": "^21.0.0",
     "@metamask/snaps-execution-environments": "^11.1.1",
     "@metamask/snaps-rpc-methods": "^17.0.0",
     "@metamask/snaps-sdk": "^11.1.1",
